+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.12 ] - 2023/11/11
+
+  *  Added new methods to SoundTouchClient: GetAudioProductToneControls, SetAudioProductToneControls.
+
 ###### [ 1.0.11 ] - 2023/11/11
 
   *  Added several new methods to SoundTouchClient: GetAudioDspControls, GetBlueToothInfo, GetProductCecHdmiControl, GetProductHdmiAssignmentControls, GetRebroadcastLatencyMode, SetAudioDspControlAudioMode.